Move Byron transaction tests to Byron suite
