+++ lib/core-integration/src/Test/Integration/Scenario/API/Shelley/Transactions.hs
@@ -89,8 +89,6 @@ import Data.Time.Utils
     ( utcTimePred, utcTimeSucc )
 import Data.Word
     ( Word32 )
-import Network.HTTP.Types.Method
-    ( Method )
 import Numeric.Natural
     ( Natural )
 import Test.Hspec
@@ -119,7 +117,6 @@ import Test.Integration.Framework.DSL
     , expectSuccess
     , faucetAmt
     , faucetUtxoAmt
-    , fixtureIcarusWallet
     , fixtureIcarusWalletAddrs
     , fixtureMultiAssetWallet
     , fixturePassphrase
@@ -135,6 +132,7 @@ import Test.Integration.Framework.DSL
     , listTransactions
     , minUTxOValue
     , oneSecond
+    , postTx
     , postWallet
     , request
     , rewardWallet
@@ -283,7 +281,7 @@ spec = describe "SHELLEY_TRANSACTIONS" $ do
         eventually "Pending tx has pendingSince field" $ do
             -- Post Tx
             let amt = (minUTxOValue :: Natural)
-            r <- postTx ctx
+            r <- postTx @n ctx
                 (wSrc, Link.createTransaction @'Shelley,fixturePassphrase)
                 wDest
                 amt
@@ -605,73 +603,6 @@ spec = describe "SHELLEY_TRANSACTIONS" $ do
             r <- request @(ApiTransaction n) ctx
                 (Link.createTransaction @'Shelley w) Default payload
             expectResponseCode HTTP.status400 r
-
-    describe "TRANS_CREATE_09 - Single Output Transaction with non-Shelley witnesses" $
-        forM_ [(fixtureRandomWallet, "Byron wallet"), (fixtureIcarusWallet, "Icarus wallet")] $
-        \(srcFixture,name) -> it name $ \ctx -> runResourceT $ do
-
-        (wByron, wShelley) <- (,) <$> srcFixture ctx <*> fixtureWallet ctx
-        addrs <- listAddresses @n ctx wShelley
-
-        let amt = minUTxOValue :: Natural
-        let destination = (addrs !! 1) ^. #id
-        let payload = Json [json|{
-                "payments": [{
-                    "address": #{destination},
-                    "amount": {
-                        "quantity": #{amt},
-                        "unit": "lovelace"
-                    }
-                }]
-            }|]
-
-        rFeeEst <- request @ApiFee ctx
-            (Link.getTransactionFee @'Byron wByron) Default payload
-        verify rFeeEst
-            [ expectSuccess
-            , expectResponseCode HTTP.status202
-            ]
-        let (Quantity feeEstMin) = getFromResponse #estimatedMin rFeeEst
-        let (Quantity feeEstMax) = getFromResponse #estimatedMax rFeeEst
-
-        r <- postTx ctx
-            (wByron, Link.createTransaction @'Byron, fixturePassphrase)
-            wShelley
-            amt
-        verify r
-            [ expectSuccess
-            , expectResponseCode HTTP.status202
-            , expectField (#amount . #getQuantity) $
-                between (feeEstMin + amt, feeEstMax + amt)
-            , expectField (#direction . #getApiT) (`shouldBe` Outgoing)
-            , expectField (#status . #getApiT) (`shouldBe` Pending)
-            ]
-
-        ra <- request @ApiByronWallet ctx (Link.getWallet @'Byron wByron) Default Empty
-        verify ra
-            [ expectSuccess
-            , expectField (#balance . #total) $
-                between
-                    ( Quantity (faucetAmt - feeEstMax - amt)
-                    , Quantity (faucetAmt - feeEstMin - amt)
-                    )
-            , expectField
-                    (#balance . #available)
-                    (.>= Quantity (faucetAmt - faucetUtxoAmt))
-            ]
-
-        eventually "wa and wb balances are as expected" $ do
-            rb <- request @ApiWallet ctx
-                (Link.getWallet @'Shelley wShelley) Default Empty
-            expectField
-                (#balance . #available)
-                (`shouldBe` Quantity (faucetAmt + amt)) rb
-
-            ra2 <- request @ApiByronWallet ctx
-                (Link.getWallet @'Byron wByron) Default Empty
-            expectField
-                (#balance . #available)
-                (`shouldBe` Quantity (faucetAmt - feeEstMax - amt)) ra2
 
     it "TRANS_ASSETS_CREATE_01 - Multi-asset balance" $ \ctx -> runResourceT $ do
         w <- fixtureMultiAssetWallet ctx
@@ -1974,10 +1905,10 @@ spec = describe "SHELLEY_TRANSACTIONS" $ do
         (wSrc, w) <- (,) <$> fixtureWallet ctx <*> emptyWallet ctx
         -- post txs
         let linkTx = (wSrc, Link.createTransaction @'Shelley, "cardano-wallet")
-        _ <- postTx ctx linkTx w minUTxOValue
+        _ <- postTx @n ctx linkTx w minUTxOValue
         verifyWalletBalance ctx w (Quantity minUTxOValue)
 
-        _ <- postTx ctx linkTx w (2 * minUTxOValue)
+        _ <- postTx @n ctx linkTx w (2 * minUTxOValue)
         verifyWalletBalance ctx w (Quantity (3 * minUTxOValue))
 
         txs <- eventually "I make sure there are exactly 2 transactions" $ do
@@ -2343,7 +2274,7 @@ spec = describe "SHELLEY_TRANSACTIONS" $ do
         (wSrc, wDest) <- (,) <$> fixtureWallet ctx <*> emptyWallet ctx
         -- post tx
         let amt = (minUTxOValue :: Natural)
-        rMkTx <- postTx ctx
+        rMkTx <- postTx @n ctx
             (wSrc, Link.createTransaction @'Shelley, "cardano-wallet")
             wDest
             amt
@@ -2397,7 +2328,7 @@ spec = describe "SHELLEY_TRANSACTIONS" $ do
         (wSrc, wDest) <- (,) <$> fixtureWallet ctx <*> emptyWallet ctx
         -- post tx
         let amt = (minUTxOValue :: Natural)
-        rMkTx <- postTx ctx
+        rMkTx <- postTx @n ctx
             (wSrc, Link.createTransaction @'Shelley, "cardano-wallet")
             wDest
             amt
@@ -2420,7 +2351,7 @@ spec = describe "SHELLEY_TRANSACTIONS" $ do
         (wSrc, wDest) <- (,) <$> fixtureWallet ctx <*> emptyWallet ctx
         -- post tx
         let amt = (minUTxOValue :: Natural)
-        rMkTx <- postTx ctx
+        rMkTx <- postTx @n ctx
             (wSrc, Link.createTransaction @'Shelley, "cardano-wallet")
             wDest
             amt
@@ -2484,7 +2415,7 @@ spec = describe "SHELLEY_TRANSACTIONS" $ do
 
         -- post transaction
         rTx <-
-            postTx ctx
+            postTx @n ctx
             (wSrc, Link.createTransaction @'Shelley, "cardano-wallet")
             wDest
             (minUTxOValue :: Natural)
@@ -2554,81 +2485,6 @@ spec = describe "SHELLEY_TRANSACTIONS" $ do
         -- yes, gone
         request @(ApiTransaction n) ctx linkDel Default Empty
             >>= expectResponseCode HTTP.status404
-
-    it "BYRON_TRANS_DELETE -\
-        \ Cannot delete tx on Byron wallet using shelley ep" $ \ctx -> runResourceT $ do
-            w <- emptyRandomWallet ctx
-            let wid = w ^. walletId
-            let txid = "3e6ec12da4414aa0781ff8afa9717ae53ee8cb4aa55d622f65bc62619a4f7b12"
-            let endpoint = "v2/wallets/" <> wid <> "/transactions/" <> txid
-            r <- request @ApiTxId ctx ("DELETE", endpoint) Default Empty
-            expectResponseCode HTTP.status404 r
-            expectErrorMessage (errMsg404NoWallet wid) r
-
-    it "BYRON_TRANS_ESTIMATE -\
-        \ Cannot estimate tx on Byron wallet using shelley ep" $ \ctx -> runResourceT $ do
-            w <- emptyRandomWallet ctx
-            let wid = w ^. walletId
-            wDest <- emptyWallet ctx
-            addr:_ <- listAddresses @n ctx wDest
-            let destination = addr ^. #id
-            let payload = Json [json|{
-                    "payments": [{
-                        "address": #{destination},
-                        "amount": {
-                            "quantity": #{minUTxOValue},
-                            "unit": "lovelace"
-                        }
-                    }]
-                }|]
-            let endpoint = "v2/wallets/" <> wid <> "/payment-fees"
-            r <- request @ApiFee ctx ("POST", endpoint) Default payload
-            expectResponseCode HTTP.status404 r
-            expectErrorMessage (errMsg404NoWallet wid) r
-
-    it "BYRON_TRANS_CREATE -\
-        \ Cannot create tx on Byron wallet using shelley ep" $ \ctx -> runResourceT $ do
-            w <- emptyRandomWallet ctx
-            let wid = w ^. walletId
-            wDest <- emptyWallet ctx
-            addr:_ <- listAddresses @n ctx wDest
-            let destination = addr ^. #id
-            let payload = Json [json|{
-                    "payments": [{
-                        "address": #{destination},
-                        "amount": {
-                            "quantity": #{minUTxOValue},
-                            "unit": "lovelace"
-                        }
-                    }],
-                    "passphrase": "cardano-wallet"
-                }|]
-            let endpoint = "v2/wallets/" <> wid <> "/transactions"
-            r <- request @(ApiTransaction n) ctx ("POST", endpoint) Default payload
-            expectResponseCode HTTP.status404 r
-            expectErrorMessage (errMsg404NoWallet wid) r
-
-    it "BYRON_TX_LIST_02 -\
-        \ Byron endpoint does not list Shelley wallet transactions" $ \ctx -> runResourceT $ do
-        w <- emptyWallet ctx
-        let wid = w ^. walletId
-        let ep = ("GET", "v2/byron-wallets/" <> wid <> "/transactions")
-        r <- request @([ApiTransaction n]) ctx ep Default Empty
-        verify r
-            [ expectResponseCode HTTP.status404
-            , expectErrorMessage (errMsg404NoWallet wid)
-            ]
-
-    it "BYRON_TX_LIST_03 -\
-        \ Shelley endpoint does not list Byron wallet transactions" $ \ctx -> runResourceT $ do
-        w <- emptyRandomWallet ctx
-        let wid = w ^. walletId
-        let ep = ("GET", "v2/wallets/" <> wid <> "/transactions")
-        r <- request @([ApiTransaction n]) ctx ep Default Empty
-        verify r
-            [ expectResponseCode HTTP.status404
-            , expectErrorMessage (errMsg404NoWallet wid)
-            ]
 
     it "SHELLEY_TX_REDEEM_01 - Can redeem rewards from self" $ \ctx -> runResourceT $ do
         (wSrc,_) <- rewardWallet ctx
@@ -2956,7 +2812,7 @@ spec = describe "SHELLEY_TRANSACTIONS" $ do
         it resource $ \ctx -> runResourceT $ do
             -- post tx
             (wSrc, wDest) <- (,) <$> fixtureWallet ctx <*> emptyWallet ctx
-            rMkTx <- postTx ctx
+            rMkTx <- postTx @n ctx
                 (wSrc, Link.createTransaction @'Shelley, "cardano-wallet")
                 wDest
                 (minUTxOValue :: Natural)
@@ -2971,30 +2827,6 @@ spec = describe "SHELLEY_TRANSACTIONS" $ do
             ra <- request @ApiTxId ctx ("DELETE", endpoint) Default Empty
             expectResponseCode HTTP.status404 ra
             expectErrorMessage (errMsg404CannotFindTx txid) ra
-
-    postTx
-        :: (MonadIO m, MonadUnliftIO m)
-        => Context
-        -> (wal, wal -> (Method, Text), Text)
-        -> ApiWallet
-        -> Natural
-        -> m (HTTP.Status, Either RequestException (ApiTransaction n))
-    postTx ctx (wSrc, postTxEndp, pass) wDest amt = do
-        addrs <- listAddresses @n ctx wDest
-        let destination = (addrs !! 1) ^. #id
-        let payload = Json [json|{
-                "payments": [{
-                    "address": #{destination},
-                    "amount": {
-                        "quantity": #{amt},
-                        "unit": "lovelace"
-                    }
-                }],
-                "passphrase": #{pass}
-            }|]
-        r <- request @(ApiTransaction n) ctx (postTxEndp wSrc) Default payload
-        expectResponseCode HTTP.status202 r
-        return r
 
     verifyWalletBalance
         :: (MonadIO m, MonadUnliftIO m)
